chore: rename "getResponse" to "executeHandlers"
