+++ src/core/utils/handleRequest.ts
@@ -3,7 +3,7 @@ import { Emitter } from 'strict-event-emitter'
 import { RequestHandler } from '../handlers/RequestHandler'
 import { LifeCycleEventsMap, SharedOptions } from '../sharedOptions'
 import { RequiredDeep } from '../typeUtils'
-import { ResponseLookupResult, getResponse } from './getResponse'
+import { HandlersExecutionResult, executeHandlers } from './executeHandlers'
 import { onUnhandledRequest } from './request/onUnhandledRequest'
 import { readResponseCookies } from './request/readResponseCookies'
 
@@ -38,7 +38,7 @@ export interface HandleRequestOptions {
    */
   onMockedResponse?(
     response: Response,
-    handler: RequiredDeep<ResponseLookupResult>,
+    handler: RequiredDeep<HandlersExecutionResult>,
   ): void
 }
 
@@ -61,7 +61,7 @@ export async function handleRequest(
 
   // Resolve a mocked response from the list of request handlers.
   const lookupResult = await until(() => {
-    return getResponse({
+    return executeHandlers({
       request,
       requestId,
       handlers,
@@ -116,7 +116,7 @@ export async function handleRequest(
   emitter.emit('request:match', { request, requestId })
 
   const requiredLookupResult =
-    lookupResult.data as RequiredDeep<ResponseLookupResult>
+    lookupResult.data as RequiredDeep<HandlersExecutionResult>
 
   const transformedResponse =
     handleRequestOptions?.transformResponse?.(response) ||
